fix issues with code and with it running
