+++ pyluba/mammotion/commands/messages/system.py
@@ -4,6 +4,7 @@ from typing import List
 from pyluba.mammotion.commands.abstract_message import AbstractMessage
 from pyluba.mammotion.commands.messages.navigation import MessageNavigation
 from pyluba.proto import luba_msg_pb2, mctrl_sys_pb2
+from pyluba.proto.mctrl_sys import RptInfoType
 from pyluba.utility.device_type import DeviceType
 
 
@@ -166,3 +167,51 @@ class MessageSystem(AbstractMessage):
         print(f"Send command==== IOT slim data Act {
             build.todev_report_cfg.act} {build}")
         return self.send_order_msg_sys(build)
+
+
+
+    def get_report_cfg(self, timeout: int = 10000, period: int = 1000, no_change_period: int = 2000):
+        mctlsys = mctrl_sys_pb2.MctlSys(
+            todev_report_cfg=mctrl_sys_pb2.report_info_cfg(
+                timeout=timeout,
+                period=period,
+                no_change_period=no_change_period,
+                count=1
+            )
+        )
+
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_CONNECT.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_RTK.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_DEV_LOCAL.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_WORK.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_DEV_STA.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_VISION_POINT.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_VIO.value
+        )
+        mctlsys.todev_report_cfg.sub.append(
+            RptInfoType.RIT_VISION_STATISTIC.value
+        )
+
+        lubaMsg = luba_msg_pb2.LubaMsg()
+        lubaMsg.msgtype = luba_msg_pb2.MSG_CMD_TYPE_EMBED_SYS
+        lubaMsg.sender = luba_msg_pb2.DEV_MOBILEAPP
+        lubaMsg.rcver = luba_msg_pb2.DEV_MAINCTL
+        lubaMsg.msgattr = luba_msg_pb2.MSG_ATTR_REQ
+        lubaMsg.seqs = 1
+        lubaMsg.version = 1
+        lubaMsg.subtype = 1
+        lubaMsg.sys.CopyFrom(mctlsys)
+        return lubaMsg.SerializeToString()
